requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bleach==6.0.0
Django==4.2.2
django-axes==6.0.5
django-summernote==0.8.20.0
mysqlclient==2.2.0
Pillow==9.5.0
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
webencodings==0.5.1
